+++ src/main/java/src/main/c4po/AddTaskCommand.java
@@ -1,3 +1,5 @@
+package src.main.c4po;
+
 public class AddTaskCommand extends Command {
 
     private final Task task;
