+++ src/main/java/src/main/c4po/BotException.java
@@ -1,3 +1,5 @@
+package src.main.c4po;
+
 public class BotException extends Exception {
     public BotException(String errorMessage) {
         super(errorMessage);
