+++ src/main/java/src/main/c4po/Deadline.java
@@ -1,3 +1,5 @@
+package src.main.c4po;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
