+++ src/main/java/src/main/c4po/Event.java
@@ -1,3 +1,5 @@
+package src.main.c4po;
+
 public class Event extends Task{
     protected String start;
     protected String end;
