+++ src/main/java/src/main/c4po/Storage.java
@@ -1,3 +1,7 @@
+package src.main.c4po;
+
+
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
