+++ src/main/java/src/main/c4po/Task.java
@@ -1,9 +1,4 @@
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Scanner;
+package src.main.c4po;
 import java.util.regex.Pattern;
 
 public class Task {
