+++ src/main/java/src/main/c4po/UnmarkCommand.java
@@ -1,30 +1,29 @@
-import java.io.IOException;
+package src.main.c4po;
+public class UnmarkCommand extends Command {
 
-public class MarkCommand extends Command {
-
-    Integer indexToMark;
-
-    public MarkCommand(Integer index) {
-        indexToMark = index;
+    private Integer indexToUnmark;
+    public UnmarkCommand(Integer index) {
+        this.indexToUnmark = index;
     }
+
+
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws BotException {
+
         try {
-            boolean success = tasks.mark(indexToMark, "mark");
+            boolean success = tasks.mark(indexToUnmark, "unmark");
             if (success) {
-                Ui.showMarkedDone();
-                //override storage file with all new tasks
+                Ui.showUnmarked();
                 storage.writeToFile(tasks);
             } else {
                 Ui.showMarkFail();
             }
-
-            Ui.print(tasks.getItem(indexToMark).getTaskInline());
-
+            Ui.print(tasks.getItem(indexToUnmark).getTaskInline());
         } catch (Exception e) {
             String markErr = "Sir! Index for toggling mark cannot be empty";
-            throw new BotException(markErr);
+            throw new BotException(e.getMessage() + markErr);
         }
+
     }
 
     @Override
